Type changed, mmr calculcation refactored, delete match fixed

+++ BL.API/BL.API.Core/Domain/Match/PlayerMatchRecord.cs
@@ -19,7 +19,7 @@ namespace BL.API.Core.Domain.Match
         public byte? Deaths { get; set; }
         public int? Score { get; set; }
         public byte? MVPs { get; set; }
-        public int? MMRChange { get; set; }
+        public double? MMRChange { get; set; }
         public byte? CalibrationIndex { get; set; }
     }
 }
